The escapeForJSON replaced data with a meaning change, which is not desired by the user. The problem is that a \n should NOT be replaced in ALL tokens with <br/>. This may not be intended by all apps, same for tabs! All other escapes you can see below.
So, in JSONProcessor.java I replaced:

aValue = aValue.replace("\"", "\\\"");
aValue = aValue.replace("\n", "<br/>");
aValue = aValue.replace("\t", "    ");

to this, which now properly is parsed by the client without errors and especially without changing the content and meaning of the data!

aValue = aValue
 .replace("\\", "\\\\")
 .replace("\"", "\\\"")
 .replace("\n", "\\n")
 .replace("\t", "\\t")
 .replace("\r", "\\r")
 .replace("\b", "\\b")
 .replace("\f", "\\f")
 ;

Please be aware that the escape for the backslash always needs to be the first one! ;-)