Log demo issue, the problem of the empty lines is already fixed.
This was the issue in Jira:
The log view in the demos is not so high, but there are various outputs which produce empty lines before or after a log line. This makes the logs not nice readable.

Please remove all empty lines from the logs and check for all existing demos.