Slideshow demo problems solved
In the jwsSlideshow.htm we show a log view but there is never content inside.We now support logging system with the demo embedded in an iframe.
Now the server keeps all presenters and viewers synchronized, if there is any problem in the demo should probably be some little error in the channels.