fixing according to reported issue: ---------------------------------------- From: Jain, Ila [mailto:ijain@informatica.com] Sent: Tue 9/10/2013 12:29 PM To: Alexander Schulze Subject: Fix required in TimeoutOutputStreamNIOWriter
Hi,

I recently started developing using jwebsocket libraries and found out if I stop Jwebsocketfactory and then restart in the same process it gives illegalstateexception from the timer. The problem is because the reference is not being set to null in Jwebsocketfactory.stop -> TimeoutOutputStreamNIOWriter.stopTimer(); in the stop timer method timer is just cancelled. Being static n private I cant set it to null in my code. I need to stop n restart the factory in the same jvm . Please help.